\name{births}
\alias{births}
\docType{data}
\title{
Birthrates throughout the day in four Hospitals
}
\description{The data on average number of births for each hour of the day for four hospitals.
}
\usage{data(births)}
\format{
 A double matrix with 24 observations on the following 4 variables.
  \describe{
    \item{\code{Hospital1}}{Average number of births for each hour of the day within Hospital 1}
    \item{\code{Hospital2}}{Average number of births for each hour of the day within Hospital 2}
    \item{\code{Hospital3}}{Average number of births for each hour of the day within Hospital 3}
    \item{\code{Hospital4}}{Average number of births for each hour of the day within Hospital 4}
    }
}
\source{To be determined}
\keyword{datasets}
