% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{caffeine}
\alias{caffeine}
\title{The Effects of Caffeine}
\format{
A double matrix with 28 observations on the following 3 variables.
\describe{
\item{Grade}{Grade of the Student, which is either 8th, 10th, or 12th}
\item{Without}{Number of items remembered after drinking Coke without Caffine}
\item{With}{Number of items remembered after drinking Coke with Caffine}
}
}
\source{
Claire Henson, Claire Rogers, and Nadia Reynolds. Always Coca-Cola.
Technical report, University Laboratory High School, Urbana, IL, 1996.
}
\usage{
caffeine
}
\description{
Henson et al. [1996] conducted an experiment to see whether caffeine has a
negative effect on short-term visual memory. High school students were
randomly chosen: 9 from eighth grade, 10 from tenth grade, and 9 from
twelfth grade. Each person was tested once after having caffeinated Coke,
and once after having decaffeinated Coke. After each drink, the person was
given ten seconds to try to memorize twenty small, common objects, then
allowed a minute to write down as many as could be remembered. The main
question of interest is whether people remembered more objects after the
Coke without caffeine than after the Coke with caffeine.
}
\keyword{datasets}
