% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{crabs}
\alias{crabs}
\title{Morphological Measurements on Leptograpsus Crabs}
\format{
This data frame contains the following columns:
\describe{
  \item{sp}{\code{species} - \code{"B"} or \code{"O"} for blue or orange.}
  \item{sex}{\code{"M"} (Male) or \code{"F"} (Female).}
  \item{index}{index \code{1:50} within each of the four groups.}
  \item{FL}{frontal lobe size (mm).}
  \item{RW}{rear width (mm).}
  \item{CL}{carapace length (mm).}
  \item{CW}{carapace width (mm).}
  \item{BD}{body depth (mm).}
}
}
\source{
Campbell, N.A. and Mahon, R.J. (1974) A multivariate study of
variation in two species of rock crab of genus \emph{Leptograpsus.}
\emph{Australian Journal of Zoology} \bold{22}, 417--425.

MASS, R-Package
}
\usage{
crabs
}
\description{
The \code{crabs} data frame has 200 rows and 8 columns, describing 5
morphological measurements on 50 crabs each of two colour forms and both
sexes, of the species \emph{Leptograpsus variegatus} collected at Fremantle,
W. Australia.
}
\references{
Venables, W. N. and Ripley, B. D. (2002) \emph{Modern Applied
Statistics with S.} Fourth edition.  Springer.
}
\keyword{datasets}
