\name{decathlon12}
\alias{decathlon12}
\docType{data}
\title{Decathlon Event Data from 2012 Olympics.}
\description{
The decathlon data set has scores on the top 26 men in the decathlon (a set of ten events) at the 2012 Olympics. The scores are the numbers of points each participant received in each event, plus each person's total points.
}
\usage{data(decathlon12)}
\format{
 A double matrix with 26 observations on the following 11 variables.
  \describe{
    \item{\code{Meter100}}{Individual point score for 100 Meter event.}
    \item{\code{LongJump}}{Individual point score for Long Jump event.}
    \item{\code{ShotPut}}{Individual point score for Shot Put event.}
    \item{\code{HighJump}}{Individual point score for High Jump event.}
    \item{\code{Meter400}}{Individual point score for 400 Meter event.}
    \item{\code{Hurdles110}}{Individual point score for 110 Hurdles event.}
    \item{\code{Discus}}{Individual point score for Discus event.}
    \item{\code{PoleVault}}{Individual point score for Pole Vault event.}
    \item{\code{Javelin}}{Individual point score for Javelin event.}
    \item{\code{Meter1500}}{Individual point score for 1500 Meter event.}
    \item{\code{Total}}{Individual total point score for events participated in.}
    }
}  									
\source{NBC's Olympic site}
\keyword{datasets}
