\name{exams}
\alias{exams}
\docType{data}
\title{Statistics Students' Scores on Exams}
\description{
The exams matrix has data on 191 statistics students, giving
their scores (out of 100) on the three midterm exams, and the final exam.
}
\usage{data(exams)}
\format{
 A double matrix with 191 observations on the following 4 variables.
  \describe{
    \item{\code{Midterm1}}{Student score on the first midterm out of 100.}
    \item{\code{Midterm2}}{Student score on the second midterm out of 100.}
    \item{\code{Midterm3}}{Student score on the third midterm out of 100.}
    \item{\code{FinalExam}}{Student score on the Final Exam out of 100.}
    }
}
\source{Data from one of Prof. John Marden's earlier classes}
\keyword{datasets}
