\name{grades}
\alias{grades}
\docType{data}
\title{Grades}
\description{
The data set contains grades of 107 students.
}
\usage{data(grades)}	
\format{
 A double matrix with 107 observations on the following 7 variables.
  \describe{
    \item{\code{Gender}}{Sex (0=Male, 1=Female)}
    \item{\code{HW}}{Student Score on all Homework.}
    \item{\code{Labs}}{Student Score on all Labs.}
    \item{\code{InClass}}{Student Score on all In Class work.}
    \item{\code{Midterms}}{Student Score on all Midterms.}
    \item{\code{Final}}{Student Score on the Final.}
    \item{\code{Total}}{Student's Total Score}
    }
}
\source{Data from one of Prof. John Marden's earlier classes}
\keyword{datasets}
