% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{histamine}
\alias{histamine}
\title{Histamine in Dogs}
\format{
A double matrix with 16 observations on the following 4 variables.
\describe{
  \item{Before}{Histamine levels (in micrograms per milliliter of blood)
                before the inoculation.}
  \item{After1}{Histamine levels (in micrograms per milliliter of blood)
                one minute after inoculation.}
  \item{After3}{Histamine levels (in micrograms per milliliter of blood)
                three minute after inoculation.}
  \item{After5}{Histamine levels (in micrograms per milliliter of blood)
                five minutes after inoculation.}
}
}
\source{
Kenny J.Morris and Robert Zeppa. Histamine-induced hypotension due
to morphine and arfonad in the dog. Journal of Surgical Research,
3(6):313-317, 1963.
}
\usage{
histamine
}
\description{
Sixteen dogs were treated with drugs to see the effects on their blood
histamine levels. The dogs were split into four groups: Two groups received
the drugmorphine, and two received the drug trimethaphan, both given
intravenously. For one group within each pair of drug groups, the dogs had
their supply of histamine depleted before treatment, while the other group
had histamine intact. (Measurements with the value "0.10" marked data that
was missing and, were filled with that value arbitrarily.)
}
\keyword{datasets}
