\name{mouths}
\alias{mouths}
\docType{data}
\title{Mouth Sizes}
\description{
Measurements were made on the size of mouths of 27 children at four ages: 8, 10, 12, and 14. The measurement is the distance from the "center of the pituitary to the pteryomaxillary fissure"  in millimeters. These data can be found in Potthoff and Roy [1964]. There are 11 girls (Sex=1) and 16 boys (Sex=0).
}
\usage{data(mouths)}
\format{
 A data frame with 27 observations on the following 5 variables.
  \describe{
    \item{\code{Age8}}{Measurement on child's month at age eight.}
    \item{\code{Age10}}{Measurement on child's month at age ten.}
    \item{\code{Age12}}{Measurement on child's month at age twelve.}
    \item{\code{Age14}}{Measurement on child's month at age fourteen.}
    \item{\code{Sex}}{Sex Coding: Girl=1 and Boys=0}
    }
}
\source{Richard F. Potthoff and S. N. Roy. A generalized multivariate analysis of variance model useful especially for growth curve problems. Biometrika, 51:313-326, 1964.
}
\keyword{datasets}
