\name{msos-package}
\alias{msos-package}
\alias{msos}
\docType{package}
\title{Datasets and Functions used in Multivariate Statistics - Old School by John Marden}
\description{Contains necessary Multivariate Analysis methods and Datasets used in STAT 571. This package was built from source materials provided by Prof. John Marden}
\details{
\tabular{ll}{
Package: \tab msos\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2014-01-22\cr
License: \tab MIT \cr
}
}
\author{John Marden [aut, cph] and James Balamuta [cre, ctb, com]

Maintainer: James Balamuta <james.balamuta@gmail.com>
}
\keyword{package}
