\name{planets}
\alias{planets}
\docType{data}
\title{Planets}
\description{
Six astronomical variables are given on each of the historical nine planets (or eight planets, plus Pluto).
}
\usage{data(planets)}			
\format{
 A double matrix with 9 observations on the following 6 variables.
  \describe{
    \item{\code{Distance}}{Average distance in millions of miles the planet is from the sun.}
    \item{\code{Day}}{The length of the planet's day in Earth days}
    \item{\code{Year}}{The length of year in Earth days}
    \item{\code{Diameter}}{The planet's diameter in miles}
    \item{\code{Temperature}}{The planet's temperature in degrees Fahrenheit}
    \item{\code{Moons}}{Number of moons}
    }
}
\source{
John W. Wright, editor. The Universal Almanac. Andrews McMeel Publishing, Kansas City, MO, 1997.
}
\keyword{datasets}
