\name{skulls}
\alias{skulls}
\docType{data}
\title{
Egyptian Skulls
}
\description{
The data concern the sizes of Egyptian skulls over time, from Thomson and Randall-MacIver [1905]. There are 30 skulls from each of five time periods, so that n = 150 all together.
}
\usage{data(skulls)}
\format{
 A double matrix with 150 observations on the following 5 variables.
  \describe{
    \item{\code{MaximalBreadth}}{Maximum length in millimeters}
    \item{\code{BasibregmaticHeight}}{Basibregmatic Height in millimeters}
    \item{\code{BasialveolarLength}}{Basialveolar Length in millimeters}
    \item{\code{NasalHeight}}{Nasal Height in millimeters}
    \item{\code{TimePeriod}}{Time groupings}
    }
}
\source{
A. Thomson and R. Randall-MacIver. Ancient Races of the Thebaid. Oxford University Press, 1905.
}
\keyword{datasets}
