\name{softdrinks}
\alias{softdrinks}
\docType{data}
\title{Soft Drinks}
\description{A data set that contains 23 peoples' ranking of 8 soft drinks: Coke, Pepsi, Sprite, 7-up, and their diet equivalents}
\usage{data(softdrinks)}
\format{
 A double matrix with 23 observations on the following 8 variables.
  \describe{
    \item{\code{Coke}}{Ranking given to Coke}
    \item{\code{Pepsi}}{Ranking given to Pepsi}
    \item{\code{7up}}{Ranking given to 7-up}
    \item{\code{Sprite}}{Ranking given to Sprite}
    \item{\code{DietCoke}}{Ranking given to Diet Coke}
    \item{\code{DietPepsi}}{Ranking given to Diet Pepsi}
    \item{\code{Diet7up}}{Ranking given to Diet 7-up}
    \item{\code{DietSprite}}{Ranking given to Diet Sprite}
    }
}
\source{Data from one of Prof. John Marden's earlier classes}
\keyword{datasets}
