\name{sportsranks}
\alias{sportsranks}
\docType{data}
\title{Sports ranking}
\description{
Louis Roussos asked n = 130 people to rank seven sports, assigning #1 to the sport they most wish to participate in, and #7 to the one they least wish to participate in. The sports are baseball, football, basketball, tennis, cycling, swimming and jogging.
}
\usage{data(sportsranks)} 					
\format{
 A double matrix with 130 observations on the following 7 variables.
  \describe{
    \item{\code{Baseball}}{Baseball's ranking out of seven sports.}
    \item{\code{Football}}{Football's ranking out of seven sports.}
    \item{\code{Basketball}}{Basketball's ranking out of seven sports.}
    \item{\code{Tennis}}{Tennis' ranking out of seven sports.}
    \item{\code{Cycling}}{Cycling's ranking out of seven sports.}
    \item{\code{Swimming}}{Swimming's ranking out of seven sports.}
    \item{\code{Jogging}}{Jogging's ranking out of seven sports.}
    }
}
\source{Data from one of Prof. John Marden's earlier classes}
\keyword{datasets}
