\name{states}
\alias{states}
\docType{data}
\title{States}
\description{
A data set containing several demographic variables on the 50 United States, plus D.C.
}
\usage{data(states)}
\format{
  A double matrix with 51 observations on the following 11 variables.
  \describe{
    \item{\code{Population}}{In thousands}
    \item{\code{PctCities}}{The percentage of the population that lives in metripolitan areas}
    \item{\code{Doctors}}{Number per 100,000 people}
    \item{\code{SchoolEnroll}}{The percentage enrollment in primary and secondary schools.}
    \item{\code{TeacherSalary}}{The average salary of primary and secondary school teachers.}
    \item{\code{CollegeEnroll}}{The percentage full-time enrollment at college}
    \item{\code{Crime}}{Violent crimes per 100,000 people}
    \item{\code{Prisoners}}{Number of people in prison per 10,000 people.}
    \item{\code{Poverty}}{Percentage of people below the poverty line.}
    \item{\code{Employment}}{Percentage of people employed}
    \item{\code{Income}}{Median household income}
  }
}
\source{
United States (1996) Statistical Abstract of the United States. Bureau of the Census. 
}
\references{http://www.census.gov/statab/www/ranks.html}
\keyword{datasets}
