% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{births}
\alias{births}
\title{Birthrates throughout the day in four Hospitals}
\format{A double matrix with 24 observations on the following 4 variables.
 \describe{
   \item{\code{Hospital1}}{Average number of births for each hour of the day within Hospital 1}
   \item{\code{Hospital2}}{Average number of births for each hour of the day within Hospital 2}
   \item{\code{Hospital3}}{Average number of births for each hour of the day within Hospital 3}
   \item{\code{Hospital4}}{Average number of births for each hour of the day within Hospital 4}
 }}
\source{
To be determined
}
\usage{
births
}
\description{
The data on average number of births for each hour of the day for four hospitals.
}
\keyword{datasets}
