% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{cars}
\alias{cars}
\title{Automobile Data from Consumer Reports}
\format{
A double matrix with 111 observations on the following 11 variables.
\describe{
 \item{Length}{Overall length, in inches, as supplied by manufacturer.}
 \item{Wheelbase}{Length of wheelbase, in inches, as supplied by
                  manufacturer.}
 \item{Width}{Width of car, in inches, as supplied by manufacturer.}
 \item{Height}{Height of car, in inches, as supplied by manufacturer}
 \item{FrontHd}{Distance between the car's head-liner and the head of
                a 5 ft. 9 in. front seat passenger, in inches, as measured
                by CU.}
 \item{RearHd}{Distance between the car's head-liner and the head of a
               5 ft 9 in. rear seat passenger, in inches, as measured by CU}
 \item{FrtLegRoom}{Maximum front leg room, in inches, as measured by CU.}
 \item{RearSeating}{Rear fore-and-aft seating room, in inches, as measured
                    by CU.}
 \item{FrtShld}{Front shoulder room, in inches, as measured by CU.}
 \item{RearShld}{Rear shoulder room, in inches, as measured by CU.}
 \item{Luggage}{Luggage Area in Car.}
}
}
\source{
Consumers' Union. Body dimensions. Consumer Reports, April 286 -
288, 1990.
}
\usage{
cars
}
\description{
The data set cars [Consumers' Union, 1990] contains 111 models of
automobile. The original data can be found in the S-Plus? [TIBCO Software
Inc., 2009] data frame cu.dimensions. In cars, the variables have been
normalized to have medians of 0 and median absolute deviations (MAD) of
1.4826 (the MAD for a N(0, 1)).
}
\keyword{datasets}
