% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{cereal}
\alias{cereal}
\title{Cereal}
\format{
A double matrix with 8 observations on the following 11 variables.
\describe{
  \item{Return}{A cereal one would come back to}
  \item{Tasty}{Tastes good}
  \item{Popular}{Popular with the entire family}
  \item{Nourishing}{Cereal is fulfilling}
  \item{NaturalFlavor}{Cereal lacks flavor additives}
  \item{Affordable}{Cereal is priced well for the content}
  \item{GoodValue}{Quantity for Price}
  \item{Crispy}{Stays crispy in milk}
  \item{Fit}{Keeps one fit}
  \item{Fun}{Fun for children}
}
}
\source{
T. K. Chakrapani and A. S. C. Ehrenberg. An alternative to factor
analysis in marketing research part 2: Between group analysis. Professional
Marketing Research Society Journal, 1:32-38, 1981.
}
\usage{
cereal
}
\description{
Chakrapani and Ehrenberg [1981] analyzed people's attitudes towards a
variety of breakfast cereals. The data matrix cereal is 8 ? 11, with rows
corresponding to eight cereals, and columns corresponding to potential
attributes about cereals. The original data consisted of the percentage of
subjects who thought the given cereal possessed the given attribute. The
present matrix has been doubly centered, so that the row means and columns
means are all zero. (The original data can be found in the S-Plus [TIBCO
Software Inc., 2009] data set cereal.attitude.)
}
\keyword{datasets}
