% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{decathlon08}
\alias{decathlon08}
\title{Decathlon Event Data from 2008 Olympics.}
\format{A double matrix with 24 observations on the following 11 variables.
\describe{
  \item{X_100meter}{Individual point score for 100 Meter event.}
  \item{LongJump}{Individual point score for Long Jump event.}
  \item{ShotPut}{Individual point score for Shot Put event.}
  \item{HighJump}{Individual point score for High Jump event.}
  \item{X_400meter}{Individual point score for 400 Meter event.}
  \item{Hurdles}{Individual point score for 110 Hurdles event.}
  \item{Discus}{Individual point score for Discus event.}
  \item{PoleVault}{Individual point score for Pole Vault event.}
  \item{Javelin}{Individual point score for Javelin event.}
  \item{X_1500meter}{Individual point score for 1500 Meter event.}
  \item{Total}{Individual total point score for events participated in.}
}}
\source{
NBC's Olympic site
}
\usage{
decathlon08
}
\description{
The decathlon data set has scores on the top 24 men in the decathlon (a set
of ten events) at the 2008 Olympics. The scores are the numbers of points
each participant received in each event, plus each person's total points.
}
\keyword{datasets}
