% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{decathlon12}
\alias{decathlon12}
\title{Decathlon Event Data from 2012 Olympics.}
\format{A double matrix with 26 observations on the following 11 variables.
\describe{
  \item{Meter100}{Individual point score for 100 Meter event.}
  \item{LongJump}{Individual point score for Long Jump event.}
  \item{ShotPut}{Individual point score for Shot Put event.}
  \item{HighJump}{Individual point score for High Jump event.}
  \item{Meter400}{Individual point score for 400 Meter event.}
  \item{Hurdles110}{Individual point score for 110 Hurdles event.}
  \item{Discus}{Individual point score for Discus event.}
  \item{PoleVault}{Individual point score for Pole Vault event.}
  \item{Javelin}{Individual point score for Javelin event.}
  \item{Meter1500}{Individual point score for 1500 Meter event.}
  \item{Total}{Individual total point score for events participated in.} }}
\source{
NBC's Olympic site
}
\usage{
decathlon12
}
\description{
The decathlon data set has scores on the top 26 men in the decathlon (a set
of ten events) at the 2012 Olympics. The scores are the numbers of points
each participant received in each event, plus each person's total points.
}
\keyword{datasets}
