% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{election}
\alias{election}
\title{Presidential Election Data}
\format{
A double matrix with 51 observations on the following 3 variables.
\describe{
\item{2000}{Results for 51 States in Year 2000}
\item{2004}{Results for 51 States in Year 2004}
\item{2008}{Results for 51 States in Year 2008}
}
}
\source{
Calculated by Prof. John Marden, data source to be announced.
}
\usage{
election
}
\description{
The data set election has the results of the first three US presidential
races of the 2000's (2000, 2004, 2008). The observations are the 50 states
plus the District of Columbia, and the values are the (D - R)/(D + R) for
each state and each year, where D is the number of votes the Democrat
received, and R is the number the Republican received.
}
\keyword{datasets}
