% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{exams}
\alias{exams}
\title{Statistics Students' Scores on Exams}
\format{
A double matrix with 191 observations on the following 4 variables.
\describe{
  \item{Midterm1}{Student score on the first midterm out of 100.}
  \item{Midterm2}{Student score on the second midterm out of 100.}
  \item{Midterm3}{Student score on the third midterm out of 100.}
  \item{FinalExam}{Student score on the Final Exam out of 100.}
}
}
\source{
Data from one of Prof. John Marden's earlier classes
}
\usage{
exams
}
\description{
The exams matrix has data on 191 statistics students, giving their scores
(out of 100) on the three midterm exams, and the final exam.
}
\keyword{datasets}
