% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillout.R
\name{fillout}
\alias{fillout}
\title{Make a square matrix}
\usage{
fillout(z)
}
\arguments{
\item{z}{A \eqn{Q \times (Q-L)}{Q x (Q-L)} matrix}
}
\value{
A square matrix \eqn{Q \times Q}{Q x Q}
}
\description{
The function fillout takes a \eqn{Q \times (Q - L)}{Q x (Q-L)} matrix \eqn{Z}
and fills it out so that it is a square matrix \eqn{Q \times Q}{Q x Q}.
}
\examples{
# Create a 3 x 2 matrix
a <- cbind(c(1, 2, 3), c(4, 5, 6))

# Creates a 3 x 3 Matrix from 3 x 2 Data
fillout(a)
}
\seealso{
\code{\link{tr}}, \code{\link{logdet}}
}
