% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{grades}
\alias{grades}
\title{Grades}
\format{
A double matrix with 107 observations on the following 7 variables.
\describe{
  \item{Gender}{Sex (0=Male, 1=Female)}
  \item{HW}{Student Score on all Homework.}
  \item{Labs}{Student Score on all Labs.}
  \item{InClass}{Student Score on all In Class work.}
  \item{Midterms}{Student Score on all Midterms.}
  \item{Final}{Student Score on the Final.}
  \item{Total}{Student's Total Score}
}
}
\source{
Data from one of Prof. John Marden's earlier classes
}
\usage{
grades
}
\description{
The data set contains grades of 107 students.
}
\keyword{datasets}
