% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdet.R
\name{logdet}
\alias{logdet}
\title{Log Determinant}
\usage{
logdet(a)
}
\arguments{
\item{a}{Square matrix (\eqn{Q \times Q}{Q x Q})}
}
\value{
A single-value double.
}
\description{
Takes the log determinant of a square matrix. Log is that of base e
sometimes referred to as ln().
}
\examples{
# Identity Matrix of size 2
logdet(diag(c(2, 2)))
}
\seealso{
\code{\link{tr}} and \code{\link{fillout}}
}
