% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{mouths}
\alias{mouths}
\title{Mouth Sizes}
\format{
A data frame with 27 observations on the following 5 variables.
\describe{
  \item{Age8}{Measurement on child's month at age eight.}
  \item{Age10}{Measurement on child's month at age ten.}
  \item{Age12}{Measurement on child's month at age twelve.}
  \item{Age14}{Measurement on child's month at age fourteen.}
  \item{Sex}{Sex Coding: Girl=1 and Boys=0}
}
}
\source{
Richard F. Potthoff and S. N. Roy. A generalized multivariate
analysis of variance model useful especially for growth curve problems.
Biometrika, 51:313-326, 1964.
}
\usage{
mouths
}
\description{
Measurements were made on the size of mouths of 27 children at four ages: 8,
10, 12, and 14. The measurement is the distance from the "center of the
pituitary to the pteryomaxillary fissure" in millimeters. These data can be
found in Potthoff and Roy [1964]. There are 11 girls (Sex=1) and 16 boys
(Sex=0).
}
\keyword{datasets}
