% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msos-internal.R
\docType{package}
\name{msos-package}
\alias{msos}
\alias{msos-package}
\title{msos: Data Sets and Functions Used in Multivariate Statistics: Old School by John Marden}
\description{
Contains necessary Multivariate Analysis methods and data sets used
in John Marden's book Multivariate Statistics: Old School (2015) <ISBN:978-1456538835>.
This is also serves as a companion package for the
STAT 571: Multivariate Analysis course at the University of Illinois
at Urbana-Champaign (UIUC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/msos}
  \item Report bugs at \url{https://github.com/coatless/msos/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{james.balamuta@gmail.com} [contributor, compiler]

Authors:
\itemize{
  \item John Marden \email{jimarden@illinois.edu} [copyright holder]
}

}
