% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msos-package.R
\docType{package}
\name{msos-package}
\alias{msos}
\alias{msos-package}
\title{msos: Data Sets and Functions Used in Multivariate Statistics: Old School by John Marden}
\description{
Multivariate Analysis methods and data sets used
    in John Marden's book Multivariate Statistics: Old School (2015) <ISBN:978-1456538835>.
    This also serves as a companion package for the 
    STAT 571: Multivariate Analysis course offered by the Department of Statistics
    at the University of Illinois at Urbana-Champaign ('UIUC').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/msos}
  \item \url{https://coatless.github.io/msos/}
  \item Report bugs at \url{https://github.com/coatless/msos/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{james.balamuta@gmail.com} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [contributor, compiler]

Authors:
\itemize{
  \item John Marden \email{jimarden@illinois.edu} [copyright holder]
}

}
\keyword{internal}
