% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{planets}
\alias{planets}
\title{Planets}
\format{A double matrix with 9 observations on the following 6 variables.
\describe{
  \item{Distance}{Average distance in millions of miles the planet is from
                 the sun.}
  \item{Day}{The length of the planet's day in Earth days}
  \item{Year}{The length of year in Earth days}
  \item{Diameter}{The planet's diameter in miles}
  \item{Temperature}{The planet's temperature in degrees Fahrenheit}
  \item{Moons}{Number of moons}
}}
\source{
John W. Wright, editor. The Universal Almanac. Andrews McMeel
Publishing, Kansas City, MO, 1997.
}
\usage{
planets
}
\description{
Six astronomical variables are given on each of the historical nine planets
(or eight planets, plus Pluto).
}
\keyword{datasets}
