% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{prostaglandin}
\alias{prostaglandin}
\title{Prostaglandin}
\format{
A double matrix with 10 observations on the following 6 variables.
\describe{
  \item{T1}{First four-hour interval measurement of prostaglandin.}
  \item{T2}{Second four-hour interval measurement of prostaglandin.}
  \item{T3}{Third four-hour interval measurement of prostaglandin.}
  \item{T4}{Fourth four-hour interval measurement of prostaglandin.}
  \item{T5}{Fifth four-hour interval measurement of prostaglandin.}
  \item{T6}{Sixth four-hour interval measurement of prostaglandin.}
}
}
\source{
J H Ware and R E Bowden. Circadian rhythm analysis when output is
collected at intervals. Biometrics, 33(3):566-571, 1977.
}
\usage{
prostaglandin
}
\description{
Data from Ware and Bowden [1977] taken at six four-hour intervals (labelled
T1 to T6) over the course of a day for 10 individuals. The measurements are
prostaglandin contents in their urine.
}
\keyword{datasets}
