% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{skulls}
\alias{skulls}
\title{Egyptian Skulls}
\format{A double matrix with 150 observations on the following 5 variables.
\describe{
  \item{MaximalBreadth}{Maximum length in millimeters}
  \item{BasibregmaticHeight}{Basibregmatic Height in millimeters}
  \item{BasialveolarLength}{Basialveolar Length in millimeters}
  \item{NasalHeight}{Nasal Height in millimeters}
  \item{TimePeriod}{Time groupings}
}}
\source{
A. Thomson and R. Randall-MacIver. Ancient Races of the Thebaid.
Oxford University Press, 1905.
}
\usage{
skulls
}
\description{
The data concern the sizes of Egyptian skulls over time, from Thomson and
Randall-MacIver [1905]. There are 30 skulls from each of five time periods,
so that n = 150 all together.
}
\keyword{datasets}
