% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{softdrinks}
\alias{softdrinks}
\title{Soft Drinks}
\format{A double matrix with 23 observations on the following 8 variables.
\describe{
  \item{Coke}{Ranking given to Coke}
  \item{Pepsi}{Ranking given to Pepsi}
  \item{7up}{Ranking given to 7-up}
  \item{Sprite}{Ranking given to Sprite}
  \item{DietCoke}{Ranking given to Diet Coke}
  \item{DietPepsi}{Ranking given to Diet Pepsi}
  \item{Diet7up}{Ranking given to Diet 7-up}
  \item{DietSprite}{Ranking given to Diet Sprite}
}}
\source{
Data from one of Prof. John Marden's earlier classes
}
\usage{
softdrinks
}
\description{
A data set that contains 23 peoples' ranking of 8 soft drinks: Coke, Pepsi,
Sprite, 7-up, and their diet equivalents
}
\keyword{datasets}
