% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-doc.R
\docType{data}
\name{sportsranks}
\alias{sportsranks}
\title{Sports ranking}
\format{
A double matrix with 130 observations on the following 7 variables.
\describe{
  \item{Baseball}{Baseball's ranking out of seven sports.}
  \item{Football}{Football's ranking out of seven sports.}
  \item{Basketball}{Basketball's ranking out of seven sports.}
  \item{Tennis}{Tennis' ranking out of seven sports.}
  \item{Cycling}{Cycling's ranking out of seven sports.}
  \item{Swimming}{Swimming's ranking out of seven sports.}
  \item{Jogging}{Jogging's ranking out of seven sports.}
}
}
\source{
Data from one of Prof. John Marden's earlier classes
}
\usage{
sportsranks
}
\description{
Louis Roussos asked n = 130 people to rank seven sports, assigning #1 to the
sport they most wish to participate in, and #7 to the one they least wish to
participate in. The sports are baseball, football, basketball, tennis,
cycling, swimming and jogging.
}
\keyword{datasets}
